Spotbugs errors.

+++ hazelcast/src/main/java/com/hazelcast/sql/impl/exec/io/AbstractInbox.java
@@ -114,6 +114,6 @@ public abstract class AbstractInbox extends AbstractMailbox implements InboundHa
     }
 
     private long getBatchSize(InboundBatch batch) {
-        return batch.getBatch().getRowCount() * rowWidth;
+        return (long) batch.getBatch().getRowCount() * rowWidth;
     }
 }
